openai_requirements.txt
nexaai
streamlit_modal
sounddevice

# OpenAI API support
openai